cmake: generate ASan+libFuzzer and UBSan+libFuzzer
In order for native fuzzing to be effective, native Lua
extensions must be built with Clang using the argument
`-fsanitize=fuzzer-no-link` and optionally other sanitizers.

At runtime, you may get an error about missing symbols:

| ./luac.so: undefined symbol: __ubsan_handle_builtin_unreachable

These are libFuzzer symbols, and you must ensure they are
available in the global namespace. Here are the two options:
preloading sanitizer and libFuzzer libraries and linking libFuzzer
into Lua runtime.

The patch implements the first aforementioned option. When luzer
is built, it attempts to generate custom ASan and UBSan shared
libraries that have libFuzzer linked in. These files are generated
using libFuzzer and an appropriate sanitizer library, and preinit
object files are stripped in static libraries; otherwise, a message
`.preinit_array section is not allowed in DSO` will prevent
building DSO. The resulting shared libraries will be called:

- asan_with_fuzzer.so
- ubsan_with_fuzzer.so

These files must be preloaded to allow use of a sanitizer without
having to modify Lua runtime. Simply LD_PRELOAD the right .so
file, and you're good to go.

The idea of the implementation was borrowed from Atheris, see
details in [1], [2] and [3].

1. https://github.com/google/atheris/commit/6769fecfc3ad41754cbf313f0e4ffe353dca315b
2. https://github.com/google/atheris/commit/6c17ee407bff81206af2bbd57c480ddd35a449ef
3. https://github.com/google/atheris/blob/cbf4ad989dcb4d3ef42152990ed89cfceb50e059/native_extension_fuzzing.md