luzer: initialize FDP metatable only once
The FDP metatable is recreating each time when we initialize the FDP
object. It increases the GC consumption and makes the code less
JIT-friendly since traces are exited by the guard on the FDP methods
objects that checked the exact pointer of the `GCobj` which is new for
any new FDP provider.

This patch fixes that by initializing the metatable only once on the
luzer library loading.