luzer: introduce a path module method
The patch introduces a method `path()` that returns a path to
shared libraries with libFuzzer and sanitizers.
`require('luzer').path` returns a table with paths to shared
libraries with sanitizers:

```
> luzer.path
---
- ubsan: ./build/luzer/lf_ubsan.so
  asan: ./build/luzer/lf_asan.so
  ubsan_cxx: ./build/luzer/lf_ubsan_cxx.so
```

Usage: LD_PRELOAD="$(luajit -e "print(require('luzer').path.asan)")" \
  luajit examples/example_basic.lua

Warning: some C applications may have a known flaws that will
trigger sanitizers. And it is sometimes useful to tell
sanitizers to instrument only a subset of the functions in your
target without modifying source files. With
`-fsanitize-coverage-allowlist=allowlist.txt` and
`-fsanitize-coverage-ignorelist=blocklist.txt`, you can specify
such a subset through the combination of an allowlist and a blocklist.

1. https://clang.llvm.org/docs/SanitizerCoverage.html#disabling-instrumentation-without-source-modification