luzer: support luacov
PUC Rio Lua and LuaJIT have `debug` library that includes a
`lua_sethook()` function [1]. The function sets the given function
as a hook. An implementation of `lua_sethook()` in PUC Rio Lua and
LuaJIT allows setting only a single function as a hook, so
simultaneously using several functions is not possible, the latest
`lua_sethook()` wins.

The `lua_sethook()` is used for instrumentation of the Lua source
code and providing feedback about code coverage to a fuzzing engine
and the same mechanism is used by the popular libraries
luacov [2] and cluacov [3], coverage analyzers for Lua scripts.
It means that it not possible to use instrumentation and code
coverage tools at the same time.

The patch implements a workaround - luzer will not set hook, and
therefore instruments a Lua code, when number of runs is equal to 1.

Note, that another possible solution is using `runner.debug_hook()`,
provided by LuaCov [4].

TODO: documentation

1. https://www.lua.org/manual/5.1/manual.html#pdf-debug.sethook
2. https://github.com/lunarmodules/luacov
3. https://github.com/lunarmodules/cluacov
4. https://lunarmodules.github.io/luacov/doc/modules/luacov.runner.html#debug_hook
5. https://github.com/lunarmodules/luacov/blob/7c2238df934c871b113bfeb0ec6c4b9130d56603/src/luacov/runner.lua#L102-L117