luzer: fix running with -jobs option
The commit d777927e538b ("luzer: fix argv building") sets `argv[0]`
to a constant value on the assumption that it is not used anywhere.
But `argv[0]` is used in running a test by multiple processes.
The patch fixes an `argv[0]` by setting it to a command and it's
first argument with path to test.

Fixes #32