luzer: fix arguments order in FDP functions
The documentation describes arguments for the similar FDP functions
`consume_strings()`, `consume_integers()` and `consume_numbers()`
in the following order: limits for generated values are first
and then a number of generated values. Fortunately, there was a bug
in `consume_numbers()` and `consume_integers()`, when first argument
was processed as a `count` (number of generated values). The patch
fixes that.

Fixes #44