luzer: add a test with bad dictionary
When luzer runs a test with broken dictionary it runs
`LLVMFuzzerRunDriver()` that prints an error message:

ParseDictionaryFile: error in line 1
		"\\200\\000" # Uses: 283

the it frees argv[] and returns. A function `free_argv()` frees
argv[] again and we observe a segmentation fault. The problem
was fixed by the previous commit. The patch adds an additional
test for the problem.

Follows up #52
Closes #65